requirements.txt
arrow
backcall
cycler
debugpy
decorator
entrypoints
ipykernel
ipython
ipython-genutils
jedi
Jinja2
jinja2-time
joblib
jupyter-client
jupyter-core
kiwisolver
make
MarkupSafe
matplotlib
matplotlib-inline
nest-asyncio
numpy
pandas
parso
pexpect
pickleshare
Pillow
prompt-toolkit
ptyprocess
Pygments
pyparsing
python-dateutil
pytz
PyYAML
pyzmq
scikit-learn
scipy
seaborn
six
sklearn
threadpoolctl
tornado
traitlets
wcwidth
